2.0

See merge request integrations/magento!104

+++ Controller/Adminhtml/Ecommerce/Delete.php
@@ -1,73 +1,52 @@
 <?php
 
+declare(strict_types=1);
+
 namespace GetResponse\GetResponseIntegration\Controller\Adminhtml\Ecommerce;
 
 use Exception;
 use GetResponse\GetResponseIntegration\Controller\Adminhtml\AbstractController;
 use GetResponse\GetResponseIntegration\Domain\GetResponse\Api\ApiClientFactory;
+use GetResponse\GetResponseIntegration\Domain\SharedKernel\Exception\IdNotFoundException;
 use GetResponse\GetResponseIntegration\Helper\Message;
+use GetResponse\GetResponseIntegration\Helper\Route;
 use GrShareCode\Shop\Command\DeleteShopCommand;
 use GrShareCode\Shop\ShopService;
 use Magento\Backend\App\Action\Context;
-use Magento\Framework\App\ResponseInterface;
-use Magento\Framework\Controller\Result\Redirect;
 
-/**
- * Class Delete
- * @package GetResponse\GetResponseIntegration\Controller\Adminhtml\Ecommerce
- */
 class Delete extends AbstractController
 {
-    const BACK_URL = 'getresponse/ecommerce/index';
-
-    /** @var ApiClientFactory */
     private $apiClientFactory;
 
-    /**
-     * @param Context $context
-     * @param ApiClientFactory $apiClientFactory
-     */
-    public function __construct(
-        Context $context,
-        ApiClientFactory $apiClientFactory
-    ) {
+    public function __construct(Context $context, ApiClientFactory $apiClientFactory)
+    {
         parent::__construct($context);
         $this->apiClientFactory = $apiClientFactory;
     }
 
-    /**
-     * @return ResponseInterface|Redirect
-     */
     public function execute()
     {
+        parent::execute();
+
+        if (!$this->isConnected()) {
+            return $this->redirectToStore(Route::ACCOUNT_INDEX_ROUTE);
+        }
+
         try {
-            $id = $this->getRequest()->getParam('id');
+            $id = $this->request->getParam('id');
 
             if (empty($id)) {
-                throw new Exception(Message::INCORRECT_SHOP);
+                throw new IdNotFoundException(Message::INCORRECT_SHOP);
             }
 
-            $service = new ShopService($this->apiClientFactory->createGetResponseApiClient());
+            $service = new ShopService(
+                $this->apiClientFactory->createGetResponseApiClient($this->scope)
+            );
             $service->deleteShop(new DeleteShopCommand($id));
 
-            $resultRedirect = $this->resultRedirectFactory->create();
-            $resultRedirect->setPath(self::BACK_URL);
-
-            return $resultRedirect;
+            return $this->redirect($this->_redirect->getRefererUrl(), Message::SHOP_DELETED);
         } catch (Exception $e) {
-            return $this->handleException($e);
+            return $this->redirect($this->_redirect->getRefererUrl(), $e->getMessage(), true);
         }
-    }
-
-    /**
-     * @param Exception $e
-     * @return Redirect
-     */
-    private function handleException(Exception $e)
-    {
-        $resultRedirect = $this->resultRedirectFactory->create();
-        $this->messageManager->addErrorMessage($e->getMessage());
-        $resultRedirect->setPath(self::BACK_URL);
-        return $resultRedirect;
     }
 }
